Remove Maven reactor setup
Separating the whole application into multiple modules joined by a Maven
Reactor build parent sounded like a good idea "back in the days".
However when developing this proved to be more trouble than it's worth.

So let's put everything back into a single root project and be done with
it.